opentelemetry_semantic_conventions

Module attribute

Semantic Attributes

The entire set of semantic attributes (or conventions) defined by the project. The resource, metric, and trace modules reference these attributes.

Constants

ASPNETCORE_DIAGNOSTICS_EXCEPTION_RESULT

ASP.NET Core exception middleware handling result

ASPNETCORE_DIAGNOSTICS_HANDLER_TYPE

Full type name of the IExceptionHandler implementation that handled the exception.

ASPNETCORE_RATE_LIMITING_POLICY

Rate limiting policy name.

ASPNETCORE_RATE_LIMITING_RESULT

Rate-limiting result, shows whether the lease was acquired or contains a rejection reason

ASPNETCORE_REQUEST_IS_UNHANDLED

Flag indicating if request was handled by the application pipeline.

ASPNETCORE_ROUTING_IS_FALLBACK

A value that indicates whether the matched route is a fallback route.

ASPNETCORE_ROUTING_MATCH_STATUS

Match result - success or failure

CLIENT_ADDRESS

Client address - domain name if available without reverse DNS lookup; otherwise, IP address or Unix domain socket name.

CLIENT_PORT

Client port number.

ERROR_TYPE

Describes a class of error the operation ended with.

EXCEPTION_ESCAPED

SHOULD be set to true if the exception event is recorded at a point where it is known that the exception is escaping the scope of the span.

EXCEPTION_MESSAGE

The exception message.

EXCEPTION_STACKTRACE

A stacktrace as a string in the natural representation for the language runtime. The representation is to be determined and documented by each language SIG.

EXCEPTION_TYPE

The type of the exception (its fully-qualified class name, if applicable). The dynamic type of the exception should be preferred over the static type in languages that support it.

HTTP_REQUEST_HEADER

HTTP request headers, <key> being the normalized HTTP Header name (lowercase), the value being the header values.

HTTP_REQUEST_METHOD

HTTP request method.

HTTP_REQUEST_METHOD_ORIGINAL

Original HTTP method sent by the client in the request line.

HTTP_REQUEST_RESEND_COUNT

The ordinal number of request resending attempt (for any reason, including redirects).

HTTP_RESPONSE_HEADER

HTTP response headers, <key> being the normalized HTTP Header name (lowercase), the value being the header values.

HTTP_RESPONSE_STATUS_CODE

HTTP response status code.

HTTP_ROUTE

The matched route, that is, the path template in the format used by the respective server framework.

JVM_GC_ACTION

Name of the garbage collector action.

JVM_GC_NAME

Name of the garbage collector.

JVM_MEMORY_POOL_NAME

Name of the memory pool.

JVM_MEMORY_TYPE

The type of memory.

JVM_THREAD_DAEMON

Whether the thread is daemon or not

JVM_THREAD_STATE

State of the thread.

NETWORK_LOCAL_ADDRESS

Local address of the network connection - IP address or Unix domain socket name.

NETWORK_LOCAL_PORT

Local port number of the network connection.

NETWORK_PEER_ADDRESS

Peer address of the network connection - IP address or Unix domain socket name.

NETWORK_PEER_PORT

Peer port number of the network connection.

NETWORK_PROTOCOL_NAME

OSI application layer or non-OSI equivalent.

NETWORK_PROTOCOL_VERSION

The actual version of the protocol used for network communication.

NETWORK_TRANSPORT

OSI transport layer or inter-process communication method.

NETWORK_TYPE

OSI network layer or non-OSI equivalent.

OTEL_SCOPE_NAME

The name of the instrumentation scope - (InstrumentationScope.Name in OTLP).

OTEL_SCOPE_VERSION

The version of the instrumentation scope - (InstrumentationScope.Version in OTLP).

OTEL_STATUS_CODE

Name of the code, either “OK” or “ERROR”. MUST NOT be set if the status code is UNSET

OTEL_STATUS_DESCRIPTION

Description of the Status if it has a value, otherwise not set.

SERVER_ADDRESS

Server domain name if available without reverse DNS lookup; otherwise, IP address or Unix domain socket name.

SERVER_PORT

Server port number.

SERVICE_NAME

Logical name of the service.

SERVICE_VERSION

The version string of the service API or implementation. The format is not defined by these conventions.

SIGNALR_CONNECTION_STATUS

SignalR HTTP connection closure status.

SIGNALR_TRANSPORT

SignalR transport type

TELEMETRY_SDK_LANGUAGE

The language of the telemetry SDK

TELEMETRY_SDK_NAME

The name of the telemetry SDK as defined above.

TELEMETRY_SDK_VERSION

The version string of the telemetry SDK.

URL_FRAGMENT

The URI fragment component

URL_FULL

Absolute URL describing a network resource according to RFC3986

URL_PATH

The URI path component

URL_QUERY

The URI query component

URL_SCHEME

The URI scheme component identifying the used protocol.

USER_AGENT_ORIGINAL

Value of the HTTP User-Agent header sent by the client.