Module opentelemetry_semantic_conventions::trace

Trace Semantic Conventions

The trace semantic conventions define a set of standardized attributes to be used in Spans.

Usage

use opentelemetry::{global, trace::Tracer as _};
use opentelemetry_semantic_conventions as semcov;

let tracer = global::tracer("my-component");
let _span = tracer
    .span_builder("span-name")
    .with_attributes(vec![
        semcov::trace::NET_PEER_NAME.string("example.org"),
        semcov::trace::NET_PEER_PORT.i64(80),
    ])
    .start(&tracer);

Constants

AWS_DYNAMODB_ATTRIBUTES_TO_GET

The value of the AttributesToGet request parameter.

AWS_DYNAMODB_ATTRIBUTE_DEFINITIONS

The JSON-serialized value of each item in the AttributeDefinitions request field.

AWS_DYNAMODB_CONSISTENT_READ

The value of the ConsistentRead request parameter.

AWS_DYNAMODB_CONSUMED_CAPACITY

The JSON-serialized value of each item in the ConsumedCapacity response field.

AWS_DYNAMODB_COUNT

The value of the Count response parameter.

AWS_DYNAMODB_EXCLUSIVE_START_TABLE

The value of the ExclusiveStartTableName request parameter.

AWS_DYNAMODB_GLOBAL_SECONDARY_INDEXES

The JSON-serialized value of each item of the GlobalSecondaryIndexes request field.

AWS_DYNAMODB_GLOBAL_SECONDARY_INDEX_UPDATES

The JSON-serialized value of each item in the the GlobalSecondaryIndexUpdates request field.

AWS_DYNAMODB_INDEX_NAME

The value of the IndexName request parameter.

AWS_DYNAMODB_ITEM_COLLECTION_METRICS

The JSON-serialized value of the ItemCollectionMetrics response field.

AWS_DYNAMODB_LIMIT

The value of the Limit request parameter.

AWS_DYNAMODB_LOCAL_SECONDARY_INDEXES

The JSON-serialized value of each item of the LocalSecondaryIndexes request field.

AWS_DYNAMODB_PROJECTION

The value of the ProjectionExpression request parameter.

AWS_DYNAMODB_PROVISIONED_READ_CAPACITY

The value of the ProvisionedThroughput.ReadCapacityUnits request parameter.

AWS_DYNAMODB_PROVISIONED_WRITE_CAPACITY

The value of the ProvisionedThroughput.WriteCapacityUnits request parameter.

AWS_DYNAMODB_SCANNED_COUNT

The value of the ScannedCount response parameter.

AWS_DYNAMODB_SCAN_FORWARD

The value of the ScanIndexForward request parameter.

AWS_DYNAMODB_SEGMENT

The value of the Segment request parameter.

AWS_DYNAMODB_SELECT

The value of the Select request parameter.

AWS_DYNAMODB_TABLE_COUNT

The the number of items in the TableNames response parameter.

AWS_DYNAMODB_TABLE_NAMES

The keys in the RequestItems object field.

AWS_DYNAMODB_TOTAL_SEGMENTS

The value of the TotalSegments request parameter.

AWS_LAMBDA_INVOKED_ARN

The full invoked ARN as provided on the Context passed to the function (Lambda-Runtime-Invoked-Function-Arn header on the /runtime/invocation/next applicable).

AWS_REQUEST_ID

The AWS request ID as returned in the response headers x-amz-request-id or x-amz-requestid.

AWS_S3_BUCKET

The S3 bucket name the request refers to. Corresponds to the --bucket parameter of the S3 API operations.

AWS_S3_COPY_SOURCE

The source object (in the form bucket/key) for the copy operation.

AWS_S3_DELETE

The delete request container that specifies the objects to be deleted.

AWS_S3_KEY

The S3 object key the request refers to. Corresponds to the --key parameter of the S3 API operations.

AWS_S3_PART_NUMBER

The part number of the part being uploaded in a multipart-upload operation. This is a positive integer between 1 and 10,000.

AWS_S3_UPLOAD_ID

Upload ID that identifies the multipart upload.

CLIENT_ADDRESS

Client address - unix domain socket name, IPv4 or IPv6 address.

CLIENT_PORT

Client port number.

CLIENT_SOCKET_ADDRESS

Immediate client peer address - unix domain socket name, IPv4 or IPv6 address.

CLIENT_SOCKET_PORT

Immediate client peer port number.

CLOUDEVENTS_EVENT_ID

The event_id uniquely identifies the event.

CLOUDEVENTS_EVENT_SOURCE

The source identifies the context in which an event happened.

CLOUDEVENTS_EVENT_SPEC_VERSION

The version of the CloudEvents specification which the event uses.

CLOUDEVENTS_EVENT_SUBJECT

The subject of the event in the context of the event producer (identified by source).

CLOUDEVENTS_EVENT_TYPE

The event_type contains a value describing the type of event related to the originating occurrence.

CODE_COLUMN

The column number in code.filepath best representing the operation. It SHOULD point within the code unit named in code.function.

CODE_FILEPATH

The source code file name that identifies the code unit as uniquely as possible (preferably an absolute file path).

CODE_FUNCTION

The method or function name, or equivalent (usually rightmost part of the code unit's name).

CODE_LINENO

The line number in code.filepath best representing the operation. It SHOULD point within the code unit named in code.function.

CODE_NAMESPACE

The "namespace" within which code.function is defined. Usually the qualified class or module name, such that code.namespace + some separator + code.function form a unique identifier for the code unit.

DB_CASSANDRA_CONSISTENCY_LEVEL

The consistency level of the query. Based on consistency values from CQL.

DB_CASSANDRA_COORDINATOR_DC

The data center of the coordinating node for a query.

DB_CASSANDRA_COORDINATOR_ID

The ID of the coordinating node for a query.

DB_CASSANDRA_IDEMPOTENCE

Whether or not the query is idempotent.

DB_CASSANDRA_PAGE_SIZE

The fetch size used for paging, i.e. how many rows will be returned at once.

DB_CASSANDRA_SPECULATIVE_EXECUTION_COUNT

The number of times a query was speculatively executed. Not set or 0 if the query was not executed speculatively.

DB_CASSANDRA_TABLE

The name of the primary table that the operation is acting upon, including the keyspace name (if applicable).

DB_CONNECTION_STRING

The connection string used to connect to the database. It is recommended to remove embedded credentials.

DB_COSMOSDB_CLIENT_ID

Unique Cosmos client instance id.

DB_COSMOSDB_CONNECTION_MODE

Cosmos client connection mode.

DB_COSMOSDB_CONTAINER

Cosmos DB container name.

DB_COSMOSDB_OPERATION_TYPE

CosmosDB Operation Type.

DB_COSMOSDB_REQUEST_CHARGE

RU consumed for that operation.

DB_COSMOSDB_REQUEST_CONTENT_LENGTH

Request payload size in bytes.

DB_COSMOSDB_STATUS_CODE

Cosmos DB status code.

DB_COSMOSDB_SUB_STATUS_CODE

Cosmos DB sub status code.

DB_JDBC_DRIVER_CLASSNAME

The fully-qualified class name of the Java Database Connectivity (JDBC) driver used to connect.

DB_MONGODB_COLLECTION

The collection being accessed within the database stated in db.name.

DB_MSSQL_INSTANCE_NAME

The Microsoft SQL Server instance name connecting to. This name is used to determine the port of a named instance.

DB_NAME

This attribute is used to report the name of the database being accessed. For commands that switch the database, this should be set to the target database (even if the command fails).

DB_OPERATION

The name of the operation being executed, e.g. the MongoDB command name such as findAndModify, or the SQL keyword.

DB_REDIS_DATABASE_INDEX

The index of the database being accessed as used in the SELECT command, provided as an integer. To be used instead of the generic db.name attribute.

DB_SQL_TABLE

The name of the primary table that the operation is acting upon, including the database name (if applicable).

DB_STATEMENT

The database statement being executed.

DB_SYSTEM

An identifier for the database management system (DBMS) product being used. See below for a list of well-known identifiers.

DB_USER

Username for accessing the database.

DESTINATION_ADDRESS

Peer address, for example IP address or UNIX socket name.

DESTINATION_DOMAIN

The domain name of the destination system.

DESTINATION_PORT

Peer port number.

ENDUSER_ID

Username or client_id extracted from the access token or Authorization header in the inbound request from outside the system.

ENDUSER_ROLE

Actual/assumed role the client is making the request under extracted from token or application security context.

ENDUSER_SCOPE

Scopes or granted authorities the client currently possesses extracted from token or application security context. The value would come from the scope associated with an OAuth 2.0 Access Token or an attribute value in a SAML 2.0 Assertion.

EVENT_DOMAIN

The domain identifies the business context for the events.

EVENT_NAME

The name identifies the event.

EXCEPTION_ESCAPED

SHOULD be set to true if the exception event is recorded at a point where it is known that the exception is escaping the scope of the span.

EXCEPTION_MESSAGE

The exception message.

EXCEPTION_STACKTRACE

A stacktrace as a string in the natural representation for the language runtime. The representation is to be determined and documented by each language SIG.

EXCEPTION_TYPE

The type of the exception (its fully-qualified class name, if applicable). The dynamic type of the exception should be preferred over the static type in languages that support it.

FAAS_COLDSTART

A boolean that is true if the serverless function is executed for the first time (aka cold-start).

FAAS_CRON

A string containing the schedule period as Cron Expression.

FAAS_DOCUMENT_COLLECTION

The name of the source on which the triggering operation was performed. For example, in Cloud Storage or S3 corresponds to the bucket name, and in Cosmos DB to the database name.

FAAS_DOCUMENT_NAME

The document name/table subjected to the operation. For example, in Cloud Storage or S3 is the name of the file, and in Cosmos DB the table name.

FAAS_DOCUMENT_OPERATION

Describes the type of the operation that was performed on the data.

FAAS_DOCUMENT_TIME

A string containing the time when the data was accessed in the ISO 8601 format expressed in UTC.

FAAS_INVOCATION_ID

The invocation ID of the current function invocation.

FAAS_INVOKED_NAME

The name of the invoked function.

FAAS_INVOKED_PROVIDER

The cloud provider of the invoked function.

FAAS_INVOKED_REGION

The cloud region of the invoked function.

FAAS_TIME

A string containing the function invocation time in the ISO 8601 format expressed in UTC.

FAAS_TRIGGER

Type of the trigger which caused this function invocation.

FEATURE_FLAG_KEY

The unique identifier of the feature flag.

FEATURE_FLAG_PROVIDER_NAME

The name of the service provider that performs the flag evaluation.

FEATURE_FLAG_VARIANT

SHOULD be a semantic identifier for a value. If one is unavailable, a stringified version of the value can be used.

GRAPHQL_DOCUMENT

The GraphQL document being executed.

GRAPHQL_OPERATION_NAME

The name of the operation being executed.

GRAPHQL_OPERATION_TYPE

The type of the operation being executed.

HTTP_METHODDeprecated

Deprecated, use http.request.method instead.

HTTP_REQUEST_BODY_SIZE

The size of the request payload body in bytes. This is the number of bytes transferred excluding headers and is often, but not always, present as the Content-Length header. For requests using transport encoding, this should be the compressed size.

HTTP_REQUEST_CONTENT_LENGTHDeprecated

Deprecated, use http.request.body.size instead.

HTTP_REQUEST_METHOD

HTTP request method.

HTTP_REQUEST_METHOD_ORIGINAL

Original HTTP method sent by the client in the request line.

HTTP_RESEND_COUNT

The ordinal number of request resending attempt (for any reason, including redirects).

HTTP_RESPONSE_BODY_SIZE

The size of the response payload body in bytes. This is the number of bytes transferred excluding headers and is often, but not always, present as the Content-Length header. For requests using transport encoding, this should be the compressed size.

HTTP_RESPONSE_CONTENT_LENGTHDeprecated

Deprecated, use http.response.body.size instead.

HTTP_RESPONSE_STATUS_CODE

HTTP response status code.

HTTP_ROUTE

The matched route (path template in the format used by the respective server framework). See note below.

HTTP_SCHEMEDeprecated

Deprecated, use url.scheme instead.

HTTP_STATUS_CODEDeprecated

Deprecated, use http.response.status_code instead.

HTTP_TARGETDeprecated

Deprecated, use url.path and url.query instead.

HTTP_URLDeprecated

Deprecated, use url.full instead.

LOG_FILE_NAME

The basename of the file.

LOG_FILE_NAME_RESOLVED

The basename of the file, with symlinks resolved.

LOG_FILE_PATH

The full path to the file.

LOG_FILE_PATH_RESOLVED

The full path to the file, with symlinks resolved.

LOG_IOSTREAM

The stream associated with the log. See below for a list of well-known values.

LOG_RECORD_UID

A unique identifier for the Log Record.

MESSAGE_COMPRESSED_SIZE

Compressed size of the message in bytes.

MESSAGE_ID

MUST be calculated as two different counters starting from 1 one for sent messages and one for received message.

MESSAGE_TYPE

Whether this is a received or sent message.

MESSAGE_UNCOMPRESSED_SIZE

Uncompressed size of the message in bytes.

MESSAGING_BATCH_MESSAGE_COUNT

The number of messages sent, received, or processed in the scope of the batching operation.

MESSAGING_CLIENT_ID

A unique identifier for the client that consumes or produces a message.

MESSAGING_DESTINATION_ANONYMOUS

A boolean that is true if the message destination is anonymous (could be unnamed or have auto-generated name).

MESSAGING_DESTINATION_NAME

The message destination name.

MESSAGING_DESTINATION_TEMPLATE

Low cardinality representation of the messaging destination name.

MESSAGING_DESTINATION_TEMPORARY

A boolean that is true if the message destination is temporary and might not exist anymore after messages are processed.

MESSAGING_KAFKA_CONSUMER_GROUP

Name of the Kafka Consumer Group that is handling the message. Only applies to consumers, not producers.

MESSAGING_KAFKA_DESTINATION_PARTITION

Partition the message is sent to.

MESSAGING_KAFKA_MESSAGE_KEY

Message keys in Kafka are used for grouping alike messages to ensure they're processed on the same partition. They differ from messaging.message.id in that they're not unique. If the key is null, the attribute MUST NOT be set.

MESSAGING_KAFKA_MESSAGE_OFFSET

The offset of a record in the corresponding Kafka partition.

MESSAGING_KAFKA_MESSAGE_TOMBSTONE

A boolean that is true if the message is a tombstone.

MESSAGING_MESSAGE_CONVERSATION_ID

The conversation ID identifying the conversation to which the message belongs, represented as a string. Sometimes called "Correlation ID".

MESSAGING_MESSAGE_ID

A value used by the messaging system as an identifier for the message, represented as a string.

MESSAGING_MESSAGE_PAYLOAD_COMPRESSED_SIZE_BYTES

The compressed size of the message payload in bytes.

MESSAGING_MESSAGE_PAYLOAD_SIZE_BYTES

The (uncompressed) size of the message payload in bytes. Also use this attribute if it is unknown whether the compressed or uncompressed payload size is reported.

MESSAGING_OPERATION

A string identifying the kind of messaging operation as defined in the Operation names section above.

MESSAGING_RABBITMQ_DESTINATION_ROUTING_KEY

RabbitMQ message routing key.

MESSAGING_ROCKETMQ_CLIENT_GROUP

Name of the RocketMQ producer/consumer group that is handling the message. The client type is identified by the SpanKind.

MESSAGING_ROCKETMQ_CONSUMPTION_MODEL

Model of message consumption. This only applies to consumer spans.

MESSAGING_ROCKETMQ_MESSAGE_DELAY_TIME_LEVEL

The delay time level for delay message, which determines the message delay time.

MESSAGING_ROCKETMQ_MESSAGE_DELIVERY_TIMESTAMP

The timestamp in milliseconds that the delay message is expected to be delivered to consumer.

MESSAGING_ROCKETMQ_MESSAGE_GROUP

It is essential for FIFO message. Messages that belong to the same message group are always processed one by one within the same consumer group.

MESSAGING_ROCKETMQ_MESSAGE_KEYS

Key(s) of message, another way to mark message besides message id.

MESSAGING_ROCKETMQ_MESSAGE_TAG

The secondary classifier of message besides topic.

MESSAGING_ROCKETMQ_MESSAGE_TYPE

Type of message.

MESSAGING_ROCKETMQ_NAMESPACE

Namespace of RocketMQ resources, resources in different namespaces are individual.

MESSAGING_SYSTEM

A string identifying the messaging system.

NETWORK_CARRIER_ICC

The ISO 3166-1 alpha-2 2-character country code associated with the mobile carrier network.

NETWORK_CARRIER_MCC

The mobile carrier country code.

NETWORK_CARRIER_MNC

The mobile carrier network code.

NETWORK_CARRIER_NAME

The name of the mobile carrier.

NETWORK_CONNECTION_SUBTYPE

This describes more details regarding the connection.type. It may be the type of cell technology connection, but it could be used for describing details about a wifi connection.

NETWORK_CONNECTION_TYPE

The internet connection type.

NETWORK_PROTOCOL_NAME

OSI Application Layer or non-OSI equivalent. The value SHOULD be normalized to lowercase.

NETWORK_PROTOCOL_VERSION

Version of the application layer protocol used. See note below.

NETWORK_TRANSPORT

OSI Transport Layer or Inter-process Communication method. The value SHOULD be normalized to lowercase.

NETWORK_TYPE

OSI Network Layer or non-OSI equivalent. The value SHOULD be normalized to lowercase.

NET_HOST_NAMEDeprecated

Deprecated, use server.address.

NET_HOST_PORTDeprecated

Deprecated, use server.port.

NET_PEER_NAMEDeprecated

Deprecated, use server.address on client spans and client.address on server spans.

NET_PEER_PORTDeprecated

Deprecated, use server.port on client spans and client.port on server spans.

NET_PROTOCOL_NAMEDeprecated

Deprecated, use network.protocol.name.

NET_PROTOCOL_VERSIONDeprecated

Deprecated, use network.protocol.version.

NET_SOCK_FAMILYDeprecated

Deprecated, use network.transport and network.type.

NET_SOCK_HOST_ADDRDeprecated

Deprecated, use server.socket.address.

NET_SOCK_HOST_PORTDeprecated

Deprecated, use server.socket.port.

NET_SOCK_PEER_ADDRDeprecated

Deprecated, use server.socket.address on client spans and client.socket.address on server spans.

NET_SOCK_PEER_NAMEDeprecated

Deprecated, use server.socket.domain on client spans.

NET_SOCK_PEER_PORTDeprecated

Deprecated, use server.socket.port on client spans and client.socket.port on server spans.

NET_TRANSPORTDeprecated

Deprecated, use network.transport.

OPENTRACING_REF_TYPE

Parent-child Reference type.

OTEL_STATUS_CODE

Name of the code, either "OK" or "ERROR". MUST NOT be set if the status code is UNSET.

OTEL_STATUS_DESCRIPTION

Description of the Status if it has a value, otherwise not set.

PEER_SERVICE

The service.name of the remote service. SHOULD be equal to the actual service.name resource attribute of the remote service if any.

POOL

Name of the memory pool.

RPC_CONNECT_RPC_ERROR_CODE

The error codes of the Connect request. Error codes are always string values.

RPC_GRPC_STATUS_CODE

The numeric status code of the gRPC request.

RPC_JSONRPC_ERROR_CODE

error.code property of response if it is an error response.

RPC_JSONRPC_ERROR_MESSAGE

error.message property of response if it is an error response.

RPC_JSONRPC_REQUEST_ID

id property of request or response. Since protocol allows id to be int, string, null or missing (for notifications), value is expected to be cast to string for simplicity. Use empty string in case of null value. Omit entirely if this is a notification.

RPC_JSONRPC_VERSION

Protocol version as in jsonrpc property of request/response. Since JSON-RPC 1.0 does not specify this, the value can be omitted.

RPC_METHOD

The name of the (logical) method being called, must be equal to the $method part in the span name.

RPC_SERVICE

The full (logical) name of the service being called, including its package name, if applicable.

RPC_SYSTEM

A string identifying the remoting system. See below for a list of well-known identifiers.

SERVER_ADDRESS

Logical server hostname, matches server FQDN if available, and IP or socket address if FQDN is not known.

SERVER_PORT

Logical server port number.

SERVER_SOCKET_ADDRESS

Physical server IP address or Unix socket address. If set from the client, should simply use the socket's peer address, and not attempt to find any actual server IP (i.e., if set from client, this may represent some proxy server instead of the logical server).

SERVER_SOCKET_DOMAIN

The domain name of an immediate peer.

SERVER_SOCKET_PORT

Physical server port.

SOURCE_ADDRESS

Source address, for example IP address or Unix socket name.

SOURCE_DOMAIN

The domain name of the source system.

SOURCE_PORT

Source port number.

THREAD_ID

Current "managed" thread ID (as opposed to OS thread ID).

THREAD_NAME

Current thread name.

TYPE

The type of memory.

URL_FRAGMENT

The URI fragment component.

URL_FULL

Absolute URL describing a network resource according to RFC3986.

URL_PATH

The URI path component.

URL_QUERY

The URI query component.

URL_SCHEME

The URI scheme component identifying the used protocol.

USER_AGENT_ORIGINAL

Value of the HTTP User-Agent header sent by the client.