Module opentelemetry_semantic_conventions::resource

Resource Semantic Conventions

The resource semantic conventions define a set of standardized attributes to be used in Resources.

Usage

use opentelemetry::sdk;
use opentelemetry_semantic_conventions as semcov;

let _tracer = opentelemetry::sdk::export::trace::stdout::new_pipeline()
    .with_trace_config(sdk::trace::config().with_resource(sdk::Resource::new(vec![
        semcov::resource::SERVICE_NAME.string("my-service"),
        semcov::resource::SERVICE_NAMESPACE.string("my-namespace"),
    ])))
    .install_simple();

Constants

AWS_ECS_CLUSTER_ARN

The ARN of an ECS cluster.

AWS_ECS_CONTAINER_ARN

The Amazon Resource Name (ARN) of an ECS container instance.

AWS_ECS_LAUNCHTYPE

The launch type for an ECS task.

AWS_ECS_TASK_ARN

The ARN of an ECS task definition.

AWS_ECS_TASK_FAMILY

The task definition family this task definition is a member of.

AWS_ECS_TASK_REVISION

The revision for this task definition.

AWS_EKS_CLUSTER_ARN

The ARN of an EKS cluster.

AWS_LOG_GROUP_ARNS

The Amazon Resource Name(s) (ARN) of the AWS log group(s).

AWS_LOG_GROUP_NAMES

The name(s) of the AWS log group(s) an application is writing to.

AWS_LOG_STREAM_ARNS

The ARN(s) of the AWS log stream(s).

AWS_LOG_STREAM_NAMES

The name(s) of the AWS log stream(s) an application is writing to.

CLOUD_ACCOUNT_ID

The cloud account ID the resource is assigned to.

CLOUD_AVAILABILITY_ZONE

Cloud regions often have multiple, isolated locations known as zones to increase availability. Availability zone represents the zone where the resource is running.

CLOUD_PLATFORM

The cloud platform in use.

CLOUD_PROVIDER

Name of the cloud provider.

CLOUD_REGION

The geographical region the resource is running. Refer to your provider's docs to see the available regions, for example AWS regions, Azure regions, or Google Cloud regions.

CONTAINER_ID

Container ID. Usually a UUID, as for example used to identify Docker containers. The UUID might be abbreviated.

CONTAINER_IMAGE_NAME

Name of the image the container was built on.

CONTAINER_IMAGE_TAG

Container image tag.

CONTAINER_NAME

Container name.

CONTAINER_RUNTIME

The container runtime managing this container.

DEPLOYMENT_ENVIRONMENT

Name of the deployment environment (aka deployment tier).

DEVICE_ID

A unique identifier representing the device.

DEVICE_MODEL_IDENTIFIER

The model identifier for the device.

DEVICE_MODEL_NAME

The marketing name for the device model.

FAAS_ID

The unique ID of the function being executed.

FAAS_INSTANCE

The execution environment ID as a string.

FAAS_MAX_MEMORY

The amount of memory available to the serverless function in MiB.

FAAS_NAME

The name of the function being executed.

FAAS_VERSION

The version string of the function being executed as defined in Version Attributes.

HOST_ARCH

The CPU architecture the host system is running on.

HOST_ID

Unique host ID. For Cloud, this must be the instance_id assigned by the cloud provider.

HOST_IMAGE_ID

VM image ID. For Cloud, this value is from the provider.

HOST_IMAGE_NAME

Name of the VM image or OS install the host was instantiated from.

HOST_IMAGE_VERSION

The version string of the VM image as defined in Version Attributes.

HOST_NAME

Name of the host. On Unix systems, it may contain what the hostname command returns, or the fully qualified hostname, or another name specified by the user.

HOST_TYPE

Type of host. For Cloud, this must be the machine type.

K8S_CLUSTER_NAME

The name of the cluster.

K8S_CONTAINER_NAME

The name of the Container in a Pod template.

K8S_CRONJOB_NAME

The name of the CronJob.

K8S_CRONJOB_UID

The UID of the CronJob.

K8S_DAEMONSET_NAME

The name of the DaemonSet.

K8S_DAEMONSET_UID

The UID of the DaemonSet.

K8S_DEPLOYMENT_NAME

The name of the Deployment.

K8S_DEPLOYMENT_UID

The UID of the Deployment.

K8S_JOB_NAME

The name of the Job.

K8S_JOB_UID

The UID of the Job.

K8S_NAMESPACE_NAME

The name of the namespace that the pod is running in.

K8S_NODE_NAME

The name of the Node.

K8S_NODE_UID

The UID of the Node.

K8S_POD_NAME

The name of the Pod.

K8S_POD_UID

The UID of the Pod.

K8S_REPLICASET_NAME

The name of the ReplicaSet.

K8S_REPLICASET_UID

The UID of the ReplicaSet.

K8S_STATEFULSET_NAME

The name of the StatefulSet.

K8S_STATEFULSET_UID

The UID of the StatefulSet.

OS_DESCRIPTION

Human readable (not intended to be parsed) OS version information, like e.g. reported by ver or lsb_release -a commands.

OS_NAME

Human readable operating system name.

OS_TYPE

The operating system type.

OS_VERSION

The version string of the operating system as defined in Version Attributes.

PROCESS_COMMAND

The command used to launch the process (i.e. the command name). On Linux based systems, can be set to the zeroth string in proc/[pid]/cmdline. On Windows, can be set to the first parameter extracted from GetCommandLineW.

PROCESS_COMMAND_ARGS

All the command arguments (including the command/executable itself) as received by the process. On Linux-based systems (and some other Unixoid systems supporting procfs), can be set according to the list of null-delimited strings extracted from proc/[pid]/cmdline. For libc-based executables, this would be the full argv vector passed to main.

PROCESS_COMMAND_LINE

The full command used to launch the process as a single string representing the full command. On Windows, can be set to the result of GetCommandLineW. Do not set this if you have to assemble it just for monitoring; use process.command_args instead.

PROCESS_EXECUTABLE_NAME

The name of the process executable. On Linux based systems, can be set to the Name in proc/[pid]/status. On Windows, can be set to the base name of GetProcessImageFileNameW.

PROCESS_EXECUTABLE_PATH

The full path to the process executable. On Linux based systems, can be set to the target of proc/[pid]/exe. On Windows, can be set to the result of GetProcessImageFileNameW.

PROCESS_OWNER

The username of the user that owns the process.

PROCESS_PID

Process identifier (PID).

PROCESS_RUNTIME_DESCRIPTION

An additional description about the runtime of the process, for example a specific vendor customization of the runtime environment.

PROCESS_RUNTIME_NAME

The name of the runtime of this process. For compiled native binaries, this SHOULD be the name of the compiler.

PROCESS_RUNTIME_VERSION

The version of the runtime of this process, as returned by the runtime without modification.

SERVICE_INSTANCE_ID

The string ID of the service instance.

SERVICE_NAME

Logical name of the service.

SERVICE_NAMESPACE

A namespace for service.name.

SERVICE_VERSION

The version string of the service API or implementation.

TELEMETRY_AUTO_VERSION

The version string of the auto instrumentation agent, if used.

TELEMETRY_SDK_LANGUAGE

The language of the telemetry SDK.

TELEMETRY_SDK_NAME

The name of the telemetry SDK as defined above.

TELEMETRY_SDK_VERSION

The version string of the telemetry SDK.

WEBENGINE_DESCRIPTION

Additional description of the web engine (e.g. detailed version and edition information).

WEBENGINE_NAME

The name of the web engine.

WEBENGINE_VERSION

The version of the web engine.