Module opentelemetry_semantic_conventions::trace

Trace Semantic Conventions

The trace semantic conventions define a set of standardized attributes to be used in Spans.

Usage

use opentelemetry::{global, trace::Tracer as _};
use opentelemetry_semantic_conventions as semcov;

let tracer = global::tracer("my-component");
let _span = tracer
    .span_builder("span-name")
    .with_attributes(vec![
        semcov::trace::NET_PEER_IP.string("10.0.0.1"),
        semcov::trace::NET_PEER_PORT.i64(80),
    ])
    .start(&tracer);

Constants

AWS_DYNAMODB_ATTRIBUTES_TO_GET

The value of the AttributesToGet request parameter.

AWS_DYNAMODB_ATTRIBUTE_DEFINITIONS

The JSON-serialized value of each item in the AttributeDefinitions request field.

AWS_DYNAMODB_CONSISTENT_READ

The value of the ConsistentRead request parameter.

AWS_DYNAMODB_CONSUMED_CAPACITY

The JSON-serialized value of each item in the ConsumedCapacity response field.

AWS_DYNAMODB_COUNT

The value of the Count response parameter.

AWS_DYNAMODB_EXCLUSIVE_START_TABLE

The value of the ExclusiveStartTableName request parameter.

AWS_DYNAMODB_GLOBAL_SECONDARY_INDEXES

The JSON-serialized value of each item of the GlobalSecondaryIndexes request field.

AWS_DYNAMODB_GLOBAL_SECONDARY_INDEX_UPDATES

The JSON-serialized value of each item in the the GlobalSecondaryIndexUpdates request field.

AWS_DYNAMODB_INDEX_NAME

The value of the IndexName request parameter.

AWS_DYNAMODB_ITEM_COLLECTION_METRICS

The JSON-serialized value of the ItemCollectionMetrics response field.

AWS_DYNAMODB_LIMIT

The value of the Limit request parameter.

AWS_DYNAMODB_LOCAL_SECONDARY_INDEXES

The JSON-serialized value of each item of the LocalSecondaryIndexes request field.

AWS_DYNAMODB_PROJECTION

The value of the ProjectionExpression request parameter.

AWS_DYNAMODB_PROVISIONED_READ_CAPACITY

The value of the ProvisionedThroughput.ReadCapacityUnits request parameter.

AWS_DYNAMODB_PROVISIONED_WRITE_CAPACITY

The value of the ProvisionedThroughput.WriteCapacityUnits request parameter.

AWS_DYNAMODB_SCANNED_COUNT

The value of the ScannedCount response parameter.

AWS_DYNAMODB_SCAN_FORWARD

The value of the ScanIndexForward request parameter.

AWS_DYNAMODB_SEGMENT

The value of the Segment request parameter.

AWS_DYNAMODB_SELECT

The value of the Select request parameter.

AWS_DYNAMODB_TABLE_COUNT

The the number of items in the TableNames response parameter.

AWS_DYNAMODB_TABLE_NAMES

The keys in the RequestItems object field.

AWS_DYNAMODB_TOTAL_SEGMENTS

The value of the TotalSegments request parameter.

CODE_FILEPATH

The source code file name that identifies the code unit as uniquely as possible (preferably an absolute file path).

CODE_FUNCTION

The method or function name, or equivalent (usually rightmost part of the code unit's name).

CODE_LINENO

The line number in code.filepath best representing the operation. It SHOULD point within the code unit named in code.function.

CODE_NAMESPACE

The "namespace" within which code.function is defined. Usually the qualified class or module name, such that code.namespace + some separator + code.function form a unique identifier for the code unit.

DB_CASSANDRA_CONSISTENCY_LEVEL

The consistency level of the query. Based on consistency values from CQL.

DB_CASSANDRA_COORDINATOR_DC

The data center of the coordinating node for a query.

DB_CASSANDRA_COORDINATOR_ID

The ID of the coordinating node for a query.

DB_CASSANDRA_IDEMPOTENCE

Whether or not the query is idempotent.

DB_CASSANDRA_KEYSPACE

The name of the keyspace being accessed. To be used instead of the generic db.name attribute.

DB_CASSANDRA_PAGE_SIZE

The fetch size used for paging, i.e. how many rows will be returned at once.

DB_CASSANDRA_SPECULATIVE_EXECUTION_COUNT

The number of times a query was speculatively executed. Not set or 0 if the query was not executed speculatively.

DB_CASSANDRA_TABLE

The name of the primary table that the operation is acting upon, including the schema name (if applicable).

DB_CONNECTION_STRING

The connection string used to connect to the database. It is recommended to remove embedded credentials.

DB_HBASE_NAMESPACE

The HBase namespace being accessed. To be used instead of the generic db.name attribute.

DB_JDBC_DRIVER_CLASSNAME

The fully-qualified class name of the Java Database Connectivity (JDBC) driver used to connect.

DB_MONGODB_COLLECTION

The collection being accessed within the database stated in db.name.

DB_MSSQL_INSTANCE_NAME

The Microsoft SQL Server instance name connecting to. This name is used to determine the port of a named instance.

DB_NAME

If no tech-specific attribute is defined, this attribute is used to report the name of the database being accessed. For commands that switch the database, this should be set to the target database (even if the command fails).

DB_OPERATION

The name of the operation being executed, e.g. the MongoDB command name such as findAndModify, or the SQL keyword.

DB_REDIS_DATABASE_INDEX

The index of the database being accessed as used in the SELECT command, provided as an integer. To be used instead of the generic db.name attribute.

DB_SQL_TABLE

The name of the primary table that the operation is acting upon, including the schema name (if applicable).

DB_STATEMENT

The database statement being executed.

DB_SYSTEM

An identifier for the database management system (DBMS) product being used. See below for a list of well-known identifiers.

DB_USER

Username for accessing the database.

ENDUSER_ID

Username or client_id extracted from the access token or Authorization header in the inbound request from outside the system.

ENDUSER_ROLE

Actual/assumed role the client is making the request under extracted from token or application security context.

ENDUSER_SCOPE

Scopes or granted authorities the client currently possesses extracted from token or application security context. The value would come from the scope associated with an OAuth 2.0 Access Token or an attribute value in a SAML 2.0 Assertion.

EXCEPTION_ESCAPED

SHOULD be set to true if the exception event is recorded at a point where it is known that the exception is escaping the scope of the span.

EXCEPTION_MESSAGE

The exception message.

EXCEPTION_STACKTRACE

A stacktrace as a string in the natural representation for the language runtime. The representation is to be determined and documented by each language SIG.

EXCEPTION_TYPE

The type of the exception (its fully-qualified class name, if applicable). The dynamic type of the exception should be preferred over the static type in languages that support it.

FAAS_COLDSTART

A boolean that is true if the serverless function is executed for the first time (aka cold-start).

FAAS_CRON

A string containing the schedule period as Cron Expression.

FAAS_DOCUMENT_COLLECTION

The name of the source on which the triggering operation was performed. For example, in Cloud Storage or S3 corresponds to the bucket name, and in Cosmos DB to the database name.

FAAS_DOCUMENT_NAME

The document name/table subjected to the operation. For example, in Cloud Storage or S3 is the name of the file, and in Cosmos DB the table name.

FAAS_DOCUMENT_OPERATION

Describes the type of the operation that was performed on the data.

FAAS_DOCUMENT_TIME

A string containing the time when the data was accessed in the ISO 8601 format expressed in UTC.

FAAS_EXECUTION

The execution ID of the current function execution.

FAAS_INVOKED_NAME

The name of the invoked function.

FAAS_INVOKED_PROVIDER

The cloud provider of the invoked function.

FAAS_INVOKED_REGION

The cloud region of the invoked function.

FAAS_TIME

A string containing the function invocation time in the ISO 8601 format expressed in UTC.

FAAS_TRIGGER

Type of the trigger on which the function is executed.

HTTP_CLIENT_IP

The IP address of the original client behind all proxies, if known (e.g. from X-Forwarded-For).

HTTP_FLAVOR

Kind of HTTP protocol used.

HTTP_HOST

The value of the HTTP host header. When the header is empty or not present, this attribute should be the same.

HTTP_METHOD

HTTP request method.

HTTP_REQUEST_CONTENT_LENGTH

The size of the request payload body in bytes. This is the number of bytes transferred excluding headers and is often, but not always, present as the Content-Length header. For requests using transport encoding, this should be the compressed size.

HTTP_REQUEST_CONTENT_LENGTH_UNCOMPRESSED

The size of the uncompressed request payload body after transport decoding. Not set if transport encoding not used.

HTTP_RESPONSE_CONTENT_LENGTH

The size of the response payload body in bytes. This is the number of bytes transferred excluding headers and is often, but not always, present as the Content-Length header. For requests using transport encoding, this should be the compressed size.

HTTP_RESPONSE_CONTENT_LENGTH_UNCOMPRESSED

The size of the uncompressed response payload body after transport decoding. Not set if transport encoding not used.

HTTP_ROUTE

The matched route (path template).

HTTP_SCHEME

The URI scheme identifying the used protocol.

HTTP_SERVER_NAME

The primary server name of the matched virtual host. This should be obtained via configuration. If no such configuration can be obtained, this attribute MUST NOT be set ( net.host.name should be used instead).

HTTP_STATUS_CODE

HTTP response status code.

HTTP_TARGET

The full request target as passed in a HTTP request line or equivalent.

HTTP_URL

Full HTTP request URL in the form scheme://host[:port]/path?query[#fragment]. Usually the fragment is not transmitted over HTTP, but if it is known, it should be included nevertheless.

HTTP_USER_AGENT

Value of the HTTP User-Agent header sent by the client.

MESSAGING_CONVERSATION_ID

The conversation ID identifying the conversation to which the message belongs, represented as a string. Sometimes called "Correlation ID".

MESSAGING_DESTINATION

The message destination name. This might be equal to the span name but is required nevertheless.

MESSAGING_DESTINATION_KIND

The kind of message destination.

MESSAGING_KAFKA_CLIENT_ID

Client Id for the Consumer or Producer that is handling the message.

MESSAGING_KAFKA_CONSUMER_GROUP

Name of the Kafka Consumer Group that is handling the message. Only applies to consumers, not producers.

MESSAGING_KAFKA_MESSAGE_KEY

Message keys in Kafka are used for grouping alike messages to ensure they're processed on the same partition. They differ from messaging.message_id in that they're not unique. If the key is null, the attribute MUST NOT be set.

MESSAGING_KAFKA_PARTITION

Partition the message is sent to.

MESSAGING_KAFKA_TOMBSTONE

A boolean that is true if the message is a tombstone.

MESSAGING_MESSAGE_ID

A value used by the messaging system as an identifier for the message, represented as a string.

MESSAGING_MESSAGE_PAYLOAD_COMPRESSED_SIZE_BYTES

The compressed size of the message payload in bytes.

MESSAGING_MESSAGE_PAYLOAD_SIZE_BYTES

The (uncompressed) size of the message payload in bytes. Also use this attribute if it is unknown whether the compressed or uncompressed payload size is reported.

MESSAGING_OPERATION

A string identifying the kind of message consumption as defined in the Operation names section above. If the operation is "send", this attribute MUST NOT be set, since the operation can be inferred from the span kind in that case.

MESSAGING_PROTOCOL

The name of the transport protocol.

MESSAGING_PROTOCOL_VERSION

The version of the transport protocol.

MESSAGING_RABBITMQ_ROUTING_KEY

RabbitMQ message routing key.

MESSAGING_SYSTEM

A string identifying the messaging system.

MESSAGING_TEMP_DESTINATION

A boolean that is true if the message destination is temporary.

MESSAGING_URL

Connection string.

NET_HOST_IP

Like net.peer.ip but for the host IP. Useful in case of a multi-IP host.

NET_HOST_NAME

Local hostname or similar, see note below.

NET_HOST_PORT

Like net.peer.port but for the host port.

NET_PEER_IP

Remote address of the peer (dotted decimal for IPv4 or RFC5952 for IPv6).

NET_PEER_NAME

Remote hostname or similar, see note below.

NET_PEER_PORT

Remote port number.

NET_TRANSPORT

Transport protocol used. See note below.

PEER_SERVICE

The service.name of the remote service. SHOULD be equal to the actual service.name resource attribute of the remote service if any.

RPC_GRPC_STATUS_CODE

The numeric status code of the gRPC request.

RPC_JSONRPC_ERROR_CODE

error.code property of response if it is an error response.

RPC_JSONRPC_ERROR_MESSAGE

error.message property of response if it is an error response.

RPC_JSONRPC_METHOD

method property from request. Unlike rpc.method, this may not relate to the actual method being called. Useful for client-side traces since client does not know what will be called on the server.

RPC_JSONRPC_REQUEST_ID

id property of request or response. Since protocol allows id to be int, string, null or missing (for notifications), value is expected to be cast to string for simplicity. Use empty string in case of null value. Omit entirely if this is a notification.

RPC_JSONRPC_VERSION

Protocol version as in jsonrpc property of request/response. Since JSON-RPC 1.0 does not specify this, the value can be omitted.

RPC_METHOD

The name of the operation corresponding to the request, as returned by the AWS SDK.

RPC_SERVICE

The name of the service to which a request is made, as returned by the AWS SDK.

RPC_SYSTEM

The value aws-api.

THREAD_ID

Current "managed" thread ID (as opposed to OS thread ID).

THREAD_NAME

Current thread name.